Modify(minjae/refactor): 일반 게시판 리팩토링 #81
